chore: search.ts DTO 적용
