+++ src/controllers/index.ts
@@ -1,2 +1,3 @@
 export * from './user';
 export * from './candies';
+export * from './search';
