+++ src/dto/search.dto/index.ts
@@ -1,0 +1,6 @@
+import mongoose from 'mongoose';
+
+export interface searchDto {
+  user_id: mongoose.Types.ObjectId;
+  item: string;
+}
