+++ src/services/index.ts
@@ -1,2 +1,3 @@
 export * from './UserService';
 export * from './CandiesService';
+export * from './SearchService';
